Tab Complete Rewrite
As it has been noted in issue #44, BungeeCord versions are having issues tab completing player names. While I have very little development skills for BungeeCord, theoretically, this should fix the issue and allow players to be tab completed based on text already given, as long as there is only 1 argument. A dev or Leoko can comment on other fixes for this, but I figured it was a quick fix, so why not. :)