Remove LogManager, remove debugSqlException, use UncheckedSQLException
* Remove LogManager. AdvancedBan/logs/latest.log made support more
difficult and confused some users. Using the server console as
it was intended, as a unified information and error stream, is
much better to work with.
* Remove debugException and debugSqlException. Rethrow as
SQLException as UncheckedSQLException instead. Once a fatal error
is detected, the problem should be signalled as clearly as possible
including with a full stacktrace.